user: test-wifi: split into rlite-wifi; clean up interface
