+++ user/libs/wpa-supplicant/common.h
